ml-service/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
textblob>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
